+++ src/vyper/mod.rs
@@ -73,7 +73,7 @@ impl Compiler {
     pub fn combined_json(
         &self,
         paths: &[PathBuf],
-        evm_version: Option<compiler_llvm_context::EVMVersion>,
+        evm_version: Option<era_compiler_common::EVMVersion>,
     ) -> anyhow::Result<CombinedJson> {
         let mut command = std::process::Command::new(self.executable.as_str());
         if let Some(evm_version) = evm_version {
@@ -97,21 +97,21 @@ impl Compiler {
             );
         }
 
-        let mut combined_json: CombinedJson =
-            compiler_common::deserialize_from_slice(output.stdout.as_slice()).map_err(|error| {
-                anyhow::anyhow!(
-                    "{} subprocess output parsing error: {}\n{}",
-                    self.executable,
-                    error,
-                    compiler_common::deserialize_from_slice::<serde_json::Value>(
-                        output.stdout.as_slice()
-                    )
-                    .map(|json| serde_json::to_string_pretty(&json).expect("Always valid"))
-                    .unwrap_or_else(
-                        |_| String::from_utf8_lossy(output.stdout.as_slice()).to_string()
-                    ),
+        let mut combined_json: CombinedJson = era_compiler_common::deserialize_from_slice(
+            output.stdout.as_slice(),
+        )
+        .map_err(|error| {
+            anyhow::anyhow!(
+                "{} subprocess output parsing error: {}\n{}",
+                self.executable,
+                error,
+                era_compiler_common::deserialize_from_slice::<serde_json::Value>(
+                    output.stdout.as_slice()
                 )
-            })?;
+                .map(|json| serde_json::to_string_pretty(&json).expect("Always valid"))
+                .unwrap_or_else(|_| String::from_utf8_lossy(output.stdout.as_slice()).to_string()),
+            )
+        })?;
         combined_json.remove_evm();
         Ok(combined_json)
     }
@@ -156,21 +156,21 @@ impl Compiler {
             );
         }
 
-        let mut output: StandardJsonOutput =
-            compiler_common::deserialize_from_slice(output.stdout.as_slice()).map_err(|error| {
-                anyhow::anyhow!(
-                    "{} subprocess output parsing error: {}\n{}",
-                    self.executable,
-                    error,
-                    compiler_common::deserialize_from_slice::<serde_json::Value>(
-                        output.stdout.as_slice()
-                    )
-                    .map(|json| serde_json::to_string_pretty(&json).expect("Always valid"))
-                    .unwrap_or_else(
-                        |_| String::from_utf8_lossy(output.stdout.as_slice()).to_string()
-                    ),
+        let mut output: StandardJsonOutput = era_compiler_common::deserialize_from_slice(
+            output.stdout.as_slice(),
+        )
+        .map_err(|error| {
+            anyhow::anyhow!(
+                "{} subprocess output parsing error: {}\n{}",
+                self.executable,
+                error,
+                era_compiler_common::deserialize_from_slice::<serde_json::Value>(
+                    output.stdout.as_slice()
                 )
-            })?;
+                .map(|json| serde_json::to_string_pretty(&json).expect("Always valid"))
+                .unwrap_or_else(|_| String::from_utf8_lossy(output.stdout.as_slice()).to_string()),
+            )
+        })?;
 
         for (full_path, source) in input.sources.into_iter() {
             let last_slash_position = full_path.rfind('/');
@@ -212,7 +212,7 @@ impl Compiler {
     pub fn lll_debug(
         &self,
         path: &Path,
-        evm_version: Option<compiler_llvm_context::EVMVersion>,
+        evm_version: Option<era_compiler_common::EVMVersion>,
         optimize: bool,
     ) -> anyhow::Result<String> {
         let mut command = std::process::Command::new(self.executable.as_str());
@@ -251,7 +251,7 @@ impl Compiler {
         &self,
         version: &semver::Version,
         mut paths: Vec<PathBuf>,
-        evm_version: Option<compiler_llvm_context::EVMVersion>,
+        evm_version: Option<era_compiler_common::EVMVersion>,
         optimize: bool,
     ) -> anyhow::Result<Project> {
         paths.sort();
@@ -324,7 +324,7 @@ impl Compiler {
         for (_path, contract) in contracts.iter() {
             source_code_hasher.update(contract.source_code().as_bytes());
         }
-        let source_code_hash: [u8; compiler_common::BYTE_LENGTH_FIELD] =
+        let source_code_hash: [u8; era_compiler_common::BYTE_LENGTH_FIELD] =
             source_code_hasher.finalize_fixed().into();
 
         let project = Project::new(version.to_owned(), source_code_hash, contracts);
@@ -338,7 +338,7 @@ impl Compiler {
     pub fn extra_output(
         &self,
         path: &Path,
-        evm_version: Option<compiler_llvm_context::EVMVersion>,
+        evm_version: Option<era_compiler_common::EVMVersion>,
         extra_output: &str,
     ) -> anyhow::Result<String> {
         let mut command = std::process::Command::new(self.executable.as_str());