+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - The `Cancun` EVM version support with temporarily unsupported transient storage
 - The `MCOPY` instruction support
 - The `--evm-version` parameter to the CLI
+- An option to disable the system request memoization
 
 ## [1.3.17] - 2024-01-30